Convert ZekesScimitars plugin from kts to kt
